feat(ui-P3): Composer/附件/反馈/追问建议全面换用官方组件
- 斜杠命令面板抽为 chat/slash-panel.tsx，接入官方 Composer
  （保留 /plan /memory /skill /dag /template，官方附件按钮/拖拽/语音位）
- 附件 pill 改用 registry attachment 组件（预览弹窗/删除/发送后展示）
- 点赞点踩改为 FeedbackAdapter + ActionBarPrimitive.FeedbackPositive/Negative
  （对接 /api/feedback，提交态高亮，submittedFeedback 驱动）
- 追问建议改为 SuggestionAdapter + 官方 ThreadFollowupSuggestions
  （读取 metadata.custom.suggestions，点击自动发送）
- thread.tsx 文案本地化（tooltip/欢迎语/按钮）
- 删除 slash-composer.tsx(321行)/chat/message-actions.tsx(146行)
